Resolve ReferenceError when importing mock-handler
